feat: 允许在通过 Uri 导航应用设置页面时保留历史记录

+++ ClassIsland/Views/SettingsWindowNew.xaml.cs
@@ -40,6 +40,7 @@ using System.IO;
 using ClassIsland.Controls;
 using Path = System.IO.Path;
 using System.Collections.ObjectModel;
+using System.Web;
 using ClassIsland.Core.Models.SettingsWindow;
 using Application = System.Windows.Application;
 using YamlDotNet.Core.Tokens;
@@ -51,6 +52,8 @@ namespace ClassIsland.Views;
 /// </summary>
 public partial class SettingsWindowNew : MyWindow
 {
+    private const string KeepHistoryParameterName = "ci_keepHistory";
+
     public SettingsNewViewModel ViewModel { get; } = new();
 
     [NotNull]
@@ -205,10 +208,13 @@ public partial class SettingsWindowNew : MyWindow
 
     private async void NavigationServiceOnLoadCompleted(object sender, NavigationEventArgs e)
     {
-        if (e.ExtraData is SettingsWindowNavigationData { IsNavigateFromSettingsWindow: true })  
+        if (e.ExtraData is SettingsWindowNavigationData { IsNavigateFromSettingsWindow: true } data)  
         {
-            // 如果是从设置导航栏导航的，那么就要清除掉返回项目
-            NavigationService.RemoveBackEntry();
+            // 如果是从设置导航栏导航的，并且没有要求保留历史记录，那么就要清除掉返回项目
+            if (!data.KeepHistory)
+            {
+                NavigationService.RemoveBackEntry();
+            }
             if (!IThemeService.IsWaitForTransientDisabled)
             {
                 await Dispatcher.Yield();
@@ -253,6 +259,7 @@ public partial class SettingsWindowNew : MyWindow
             return;
         }
         Logger.LogTrace("开始导航");
+        ViewModel.IsPopupOpen = false;
         ViewModel.IsNavigating = true;
         if (ViewModel.IsViewCompressed)
         {
@@ -260,6 +267,8 @@ public partial class SettingsWindowNew : MyWindow
         }
         ViewModel.SelectedPageInfo = info;
 
+        var uriQuery = HttpUtility.ParseQueryString(uri?.Query ?? "");
+        var keepHistory = uriQuery[KeepHistoryParameterName] == "true";
         var child = LoadingAsyncBox.LoadingView as LoadingMask;
         child?.StartFakeLoading();
         if (SettingsService.Settings.ShowEchoCaveWhenSettingsPageLoading)
@@ -277,10 +286,16 @@ public partial class SettingsWindowNew : MyWindow
         ViewModel.IsDrawerOpen = false;
         ViewModel.DrawerContent = null;
         // 进行导航
-        NavigationService.RemoveBackEntry();
-        NavigationService.Navigate(page, new SettingsWindowNavigationData(true, uri != null, uri));
+        if (!keepHistory)
+        {
+            NavigationService.RemoveBackEntry();
+        }
+        NavigationService.Navigate(page, new SettingsWindowNavigationData(true, uri != null, uri, keepHistory));
         //ViewModel.FrameContent;
-        NavigationService.RemoveBackEntry();
+        if (!keepHistory)
+        {
+            NavigationService.RemoveBackEntry();
+        }
     }
 
     private SettingsPageBase? GetPage(string? id)
